backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pydantic==1.10.15
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
motor==3.3.2
python-dotenv==1.0.1
